+++ src/lib.rs
@@ -35,6 +35,7 @@ pub mod decoder;
 pub mod encoder;
 pub mod header;
 pub mod options;
+pub mod psy;
 pub mod requant;
 pub mod synth;
 pub mod tables;